pod/brain/requirements.txt
RPi.GPIO==0.7.1
click==8.1.8
Flask==3.1.0